Moved InstructionMaps into separate file - Added op code descriptive name map
